refactor(KlaviyoSwift): extract transforms/validation out of the reducer into focused files (MAGE-900)
- Move Profile.updateProfileWithProperties out of KlaviyoState.swift onto the Profile model
- Move Event enrichment into EventAPIExtension.swift (updateEventWithIdentifiers), taking identity
  primitives instead of inout KlaviyoState
- Move String.isNotEmptyOrSame next to trimWhiteSpaceOrReturnNilIfEmpty in ProfileAPIExtension.swift
- Keep all three in KlaviyoSwift (business logic, single-module use) rather than KlaviyoCore;
  the Phase 3 win here is shrinking the reducer via extraction into focused files, not module relocation
- Add KlaviyoSwiftTests for the Event and String helpers
